+++ src/warnings.ml
@@ -38,6 +38,8 @@ type kind =
   | Predicate_symbol_expected of string
   | Function_symbol_expected of string
   | Pattern_not_exhaustive of string
+  | Pattern_guard_not_exhaustive of string
+  | Ambiguous_pattern
 
 type error = location * kind
 
@@ -129,6 +131,12 @@ let pp_kind ppf = function
         "This pattern-matching is not exhaustive.@\n\
          Here is an example of a case that is not matched:@\n\
         \  %s" p
+  | Pattern_guard_not_exhaustive p ->
+      pf ppf
+        "This pattern-matching may not be exhaustive because of the guard.@\n\
+         Here is an example of a case that may not be matched:@\n\
+        \  %s" p
+  | Ambiguous_pattern -> pf ppf "Ambiguous or-pattern under guard"
 
 let styled_list l pp = List.fold_left (fun acc x -> styled x acc) pp l